+++ lib/algolia.dart
@@ -7,25 +7,16 @@ import 'package:http/http.dart' as http;
 import 'package:uuid/uuid.dart';
 
 part 'src/algolia.dart';
-
-part 'src/index_reference.dart';
-
-part 'src/index_settings.dart';
-
-part 'src/index_snapshot.dart';
-
-part 'src/query.dart';
-
-part 'src/query_snapshot.dart';
-
-part 'src/object_snapshot.dart';
-
-part 'src/object_reference.dart';
-
 part 'src/batch.dart';
-
-part 'src/task.dart';
-
 part 'src/error.dart';
-
+part 'src/event_snapshot.dart';
+part 'src/facet_value_snapshot.dart';
+part 'src/index_reference.dart';
+part 'src/index_settings.dart';
+part 'src/index_snapshot.dart';
+part 'src/object_reference.dart';
+part 'src/object_snapshot.dart';
+part 'src/query.dart';
+part 'src/query_snapshot.dart';
+part 'src/task.dart';
 part 'src/util/json_encode.dart';